Fix build problem for Gettext
Dropped useless calls to Xfce.textdomain(). It is only necessary once in the
main function. This worked around a problem to compile the program where
GETTEXT_PACKAGE is not declared before gi18n-lib.h is included. The fix in this
commit is to current it in Makefile.am in the INCLUDES.
Changed default debug level to yes instead of full in autogen.sh.


(Old svn revision: 7517)