Implement GTK2-style progress bar + further fixes
This uses a Gtk.CellRendererProgress inside
a treeview to imitate the looks of progress
bars on GTK+2. The cell background is derived
from the bar's toplevel.

In addition, CSS styles for coloring the search
entry in the DB browser and defining the padding
on the main playback buttons are applied.

Finally, the GLib search timeout in the DB
browser is being reset after triggering which
fixes a "source ID was not found when attempting
to remove it" warning on the console.